Required Libraries
I do NOT own these libraries. Credit is given inside the source. These are not modified (Gdip_All.ahk has Gdip_ImageSearch() added to it, but no code is modified). I am uploading them only to ensure that they are always available and my library may always be used to it's full extent. These libraries are not subject to the licensing given to this branch; their licensing (or lack thereof) will be stated in their source.